Implement contig information output to database
Created Node.db_values() and modified collate.py to output that
information to the database upon finishing parsing the layout
information for that node.

Next up is implementing similar behavior for edges, clusters, and
components. And calculating the N50 of the graph (which I swear I'll
get around to eventually).