Implement component information output to database
I decided to break the "pattern" here and just build up the values
for a component's entry in the database while parsing .xdot
information. This way is more efficient then sending data over
to the Component objects, just to have the Component send it back
immediately after.

The databases are now ready for use in the Javascript side of
AsmViz! That's pretty great.