Add safe file removal (prevent race conditions)
Prevents the race condition of some other process removing the
file in question before we get around to doing it. In this case,
the error is ignored (since we were going to remove the file
anyway). And if the error is something different (like, the file
in question was magically turned into a directory or something),
then we just expose that error to the user, because it's not our
fault anymore. (Not that it was to begin with, but such is life.)