Fix file I/O race conditions -- resolves #26
Uses os.fdopen() and os.open(), with a set of flags that we alter
depending on whether or not the user set the -w argument to
collate.py, to open files to which we write. This prevents
overwriting files created after check_file_existence() but before
calling os.open().

Note that we explicitly mix the patterns of "look before you leap"
and "ask forgiveness rather than permission." This allows us to
eliminate most cases in the Look-Before-You-Leap phase (in our case,
check_file_existence()) and stave off remaining race condition
problems in the Ask-Forgiveness-Rather-Than-Permission phase (in
our case, using os.fdopen() and os.open()).

(Plus, check_file_existence() generates nice-looking error messages
for the users.)

This finally resolves the race condition problems I spent a lot of
today working on fixing. I should be (finally) free to focus on
more interesting things than file-system stuff tomorrow. Some
preliminary TODOs are updating our version of Cytoscape.js, updating
our version of GraphViz, and removing wildcard imports in Python;
this should be followed by using sql.js in the Javascript side of
things. But more on that tomorrow.