Re-generate test data .db files
To contain the features I added today, like N50 calculation and
a fixed total_bp_length field for LastGraph files.

One funny coincidence -- until I implemented N50 calculation, I was
using "100" as a placeholder value for that. Simultaneously, GML
contigs are all currently treated (due to ambiguity in the file
specification) as having a length of 100 bp. So the actual N50
of the file is 100 bp!

That's why Git thinks rf_oriented.db is a renamed version of the
older database from yesterday for RF_oriented -- they contain
the exact same data, but today's database was actually generated
calculating the N50 instead of using an accidentally-perfectly-
correct hardcoded placeholder value. Huh.