Update README to discuss loading .db files
I'm fairly sure this counts as "putting the cart before the horse,"
but I wanted to get a start on cleaning up the README to discuss
.db instead of .xdot files. I figure it's sort of similar to a
project proposal document, but written in present-tense instead of
future-tense.

NOTE that, at least until we get full functionality with .db files
implemented in asmviz_viewer, I'm going to maintain .xdot upload
functionality. Maybe (key word: maybe) we'll remove that
functionality later (I'm thinking we could, say, store it in another
file, or maybe even publish it as a separate open source project
for opening arbitrary .xdot files in Cytoscape.js), but for now I
think it's important to preserve some sort of functionality for the
time being.

(And if we decide to keep both .db and .xdot parsing in xdot2cy.js,
then we can update the README to talk about both files, also.)