Remove test asm graphs/xdot files from repo
Having to commit every time I modified these files has gotten to
be a bit annoying. Plus, our switch over to using .db files has
made the need to record this stuff a lot less pertinent -- I guess
I might make a testdb/ directory in the repository in the future,
but for the time being I think we're fine to just not store this
data in the repository.

Also, I used git rm -r --cached, so don't worry -- all of this data
is still stored on my computer. And most of it is publicly available
anyway, so we should be fine.