Change (un)collapse button icon on (un)collapse
I also switched from using the "ui-icon-squaresmall-minus" icon
to the more visible "ui-icon-minus" icon. I think it looks a bit
nicer.

So, when the button says "Uncollapse," the icon displayed is a
plus sign. And when the button says "Collapse," the icon displayed
is a minus sign. It works out well!