Implement entirely relative layout -- resolve #7
We use a wrapper div with "display: table" CSS to contain the
controls div (taking up the minimum required height), and allocate
the remaining page height to the cy div. This works surprisingly
well.

A few notes:

-I also fixed the tabbing in asmviz_viewer.html. I think I started
writing the page before I fixed the indenting specifications for
my vim setup for HTML, so there was a strange mix of tab characters
and 4-space tabs. Switched everything to unambiguously just 4-space
tabs.
-It turns out that div order matters a lot -- which is why I swapped
the "code positions" of the controls and cy divs.
-I tested this on my Android phone (HTC One VX -- see specs in issue
7 comments) and it seemed to work fine, but if I get a chance I'd
like to test this on a tablet (iPad?) next week. Not very high
priority at all, though, so I'd say it definitely isn't a pressing
issue.