Add "connected component selector" spinner
To get it placed in the proper place in the control group, I had
to move the control group initialization down a bit. I dunno why
that has an effect, but I guess it does? Worth looking into, I
suppose.

Also, I defined the min/max values of the spinner based on the
number of connected components given in the "assembly" table in the
.db file (valid vals. are [1, n], where n is the number of
connected components).

I guess what's next is adding a label/etc to indicate the function
of the spinner, and then drawing the selected component accordingly?