Tentatively add auto-db close on unloading page
This might mess things up, since some browsers (Chromium included,
apparently) retain some parts of the page state after backing out
of and then going forward to a page. And closing a .db file might
result in queries not being able to be made.

Like I mention in the code, I'm really not familiar with whether or
not memory leaks are a thing we have to worry about in Javascript.
I'll look into it, but for now I want to focus on implementing
functionality without worrying about the small(-ish) details like
this.