Add auto-changing of (un)collapse button
If the user manually collapses every compound node in the graph,
or manually uncollapses every collapsed compound node in the graph,
and the button's text now reflects a useless action (e.g. for the
first case, it says "Collapse All ..." when there are no node groups
left to collapse, or for the second case, it says "Uncollapse All"
when there are no collapsed node groups left to uncollapse), then
the button text/icon will automatically change to indicate this and
offer a more suitable action.

NOTE that I think this makes the collapseAll() button text changing
redundant, so I can probably remove that in a bit. Also, I should
probably abstract changing (un)collapseButton text+icon to a sep.
function.