Fix statistics re scaffolds vs contigs (fix #40)
Since GraphML files don't have node lengths given (...as far as I
can tell, at least), we replace the "total node length" and
"N50" field values with "N/A" for GraphML files.

I also added a <p> in the infoDialog explaining this.

Note that if/when we add support for parsing more assembly graph
input types, we'll have to adjust:

-The logic in Python that determines whether or not to record
 N50/total node length
-The HTML page (infoDialog) to add that filetype to the description,
 so users can tell easily "oh this is a [scaffold|contig] assembly,
 which is why I [cannot|can] see those particular statistics"

But that's pretty much it! The JS side of things detects NULLs for
N50 and total node length and replaces them with N/As accordingly.
Pretty cool!