Replace sql.js version with memory growth ver
This should fix the occasional problems I was getting with memory
usage when keeping a database open for a while.

See https://github.com/kripken/sql.js/issues/113 for more info
on how/why this works.

I doubt this will be a cure-all solution, but it should at least
rectify some of our prior problems.