Give NULL length for scaffolds in .db (fix #61)
We assign all nodes, even lengthless scaffolds, a bp attribute --
but for scaffolds it's only used for positioning and not used for
actual biological information. When we create a .db file, we've
already laid out the nodes in GraphViz (so we have width/height
attributes in addition to node length attributes).