Switch version of sql.js back to default
Using normally-compiled sql.js is faster (due to some optimizations)
than using sql-memory-growth.js, but raises the risk of running out
of memory when processing large .db files (generally > 100 MB or so,
from what I've seen).

Now that I fixed the error of not freeing statements, this problem
should be rectified. We should be able to load .db files under 100
MB or so fine, without any memory errors.

If we get to the point where we'd like more memory to be available,
we can split up the database files by component or something and
then submit XML HTTP Requests to load each .db file. But I think
for now this should be fine, and the desktop Cytoscape solution
should cover that possibility.