Disable "copy DNA" btn when >= 1 node has no DNA
This works well, and allows for very flexible support of GraphML,
LastGraph, and future types of files. I figure disabling the button
when no DNA sequence is present for a given node in the selection
is entirely fair. I guess we could give a more concrete message
explaining this in the future, but for now this is fine.