Implement (lightweight) DNA copying (close #62)
Alright, pretty cool! It could probably be faster (querying the
DB for multiple values at once, maybe?), but I think it's alright
as is.

We can very easily modify copySelectedNodeDNA() to output more
node info or change the text formatting to make it BLAST-compatible,
etc. Look into that (#45), definitely -- it would be a cool feature.