Add functional progress bar (resolve #59)
Woo! This was annoying. Reconciling JavaScript's single-threaded
nature with timeouts, etc. wasn't the most fun thing to do.

As an aside, I really should integrate edge/cluster count into the
progress bar calculations in the future. (The effect of rendering
edges is very clear on RF_oriented's 1st largest component, but it's
not *that* bad. However, for today, this is good. Really cool.