Decrease timeout on progress bar updates
This makes the loading process a bit faster.

I guess the "risk" of using 0-length timeouts is that theoretically
the event in question could be delayed in favor of running what's
in the timeout, since the gap in which the UI would be updated is
so small. However, with something like a progress bar (where we
generally update the UI stuff many, many times), probability is on
our side, so this is a safe thing to do.