rm redundant code from pattern finding (fix #79)
This is kind of intended as a brief commit to prepare things for
implementing explicit structural pattern precedence (#77).

I'd like to fix any issues with the existing code now (while I can
verify that the fixes don't cause any problems by comparing
checksums on the output .db files) before implementing a feature
that's going to change the patterns detected in a lot of files.