Update demo salmonella info to be correct
I created a variant of the LastGraph file a while ago with some
nodes/edges added to test various structural things out. I removed
those nodes/edges when I created the demo, but I forgot to change
the node/edge statistics for the file "overview" in the selection
mechanism accordingly. Doing that now.