Add node group precedence (resolve #77)
Right now it's Bubbles > Ropes > Cycles > Chains. I can work
on implementing user-configurable precedence in the future (see #87)
but for now this should be fine, I think.

Been busy with school/etc so it took a while to get this released.
Hopefully I'll have more time to work on the next few issues.