Add PNG export for graph view (resolve #75)
I should also update the demo with the new xdot2cy.js and HTML stuff
(the HTML changes arise due to adding in a button, which messes with
the UI stuff that is declared via jQuery in the HTML file).

Also, I changed the collapse/uncollapse button text to shorten it a
bit. Ideally we'd have minimal descriptions for each button, but we
want to balance having a clean-looking UI with having clear
directions for what each button does. (We can work on UI stuff more
as we get closer to releasing this.)