Improve filetype detection; prep for GFA support
This commit adds case-insensitive support for filetype extension
detection. It also adds an error that is raised when the user passes
an input file of an invalid type.

This should make working with different filetypes easier, although
there's still some work that needs to be done on that front
(see #102, see #89).