Add support for GFA input files (resolve #89)
This is fairly bare-bones support as of now (we extract segment and
link information, and only the required information at that) --
there's more to be done here, I think. (Plus we need to handle
weird cases, like odd loops (#105) and lack of DNA (#106).)

However, for now, we should be able to get started looking at
sample GFA files.