Fix self-implying edge bug (resolve #105)
After hearing some people's opinions on what to do in this case,
I went with the option of just creating a single edge in these
cases.

As I add more specialized behavior for certain filetypes, I've begun
thinking about generalizing some of this behavior (see #108).
Could be useful to work on when I get some more time, but for now
that's a fairly low-priority issue.