Better explain node percentages (fix #103)
This makes it a lot clearer what's going on.

However, we could take this further: see #110.
this a step further and actually modify the HTML table headers/etc.
from JS, which would 1) make changing this text easier (I wouldn't
have to change the HTML of both the demo and local versions)
and 2) would allow us to do cool stuff like label the node count
field as "positive node count" or something. See #110.