Encapsulate finalized DNA export code (fix #112)
It's worth noting that the copy-and-paste text area looks a bit
odd now, since the font within that area isn't necessarily
monospace (and, therefore, even using cols="70" in the HTML won't
allow each line of DNA to fit).

However, we can address whether or not to rm the copy-and-paste
dialog stuff in the future. For now, this is good.