Now, scale nodes linearly to width and height
As opposed to scaling them linearly to the node area.

Per Dr. Pop's advice, the previous way I calculated node dimensions
involved growing them proportional to the sqrt of the log of the
node size (in either nt or bp): that is, growing the area of the
node, rather than its width or height. In an effort to make node
scaling a bit more drastic, I'm just growing nodes proportional to
the log of the node size now.

From testing this on the E. coli velvet data, this makes it a lot
easier to distinguish large and small nodes. It's still somewhat
difficult to tell apart nodes with similar log10 values (e.g.
44,000 bp vs. 130,000 bp -- I remember seeing two close nodes with
these values in the E. coli visualization), but I think that actually
makes sense, all things considered. We'd face that sort of problem
with a relative scaling method also.

...What else to say at 5 in the morning. I expect this isn't the
last time I'll have to make some changes to the scaling system, but
for now I am happy with the way things look. I'd like to get some
more feedback, of course, but I think there are probably more
pressing issues for AsmViz right now.

Wow this commit message is way too long, sorry to anyone reading this