FASTA export auto-downloads (temporary approach)
Need to test efficiency/scaling potential of this further (#114).
Basically, it might vary with browsers/systems due to the inherent
limitations of a client-side system like the current AsmViz viewer.

This is still a lot more efficient than the "copy to clipboard"
thing, but it could be better -- selecting, e.g., all of component
1 of the E. coli assembly causes this to just fail, as far as I can
tell. If we can get a better sense of the limits on this, then that
can inform the decisions we make on how to export this stuff (we
could maybe, for example, export "pointers" to the FASTA info or
something similar would allow users to get the DNA from the assembly
graph on their system).

Also should get this so that it changes the name of the file
being downloaded, hopefully (right now the file is just initially
named "download" on my system).