Add G/C content calculator
I still need to actually make use of it (calculate it for all
forward and [for LastGraph] reverse DNA strands, and then save the
results in the output .db), though.

Tested it with a few use cases -- seems to work fine. It's a pretty
simple function so hopefully there isn't a ton of room to mess it
up.