Require explicit DNA for GFA seqs (resolve #106)
Probably the easiest solution to this dilemma for now. If there
turns out to be some other standard for passing seq length in GFA,
then we can allow that to be used in lieu of an explicit DNA
sequence -- for now, though, I think this approach is fair.

You could say "well you don't require DNA sequences for Bambus 3
GraphML files" -- which is true, but those files have length
information included. And they probably could have DNA included
also in auxiliary files.