Remove some unnecessary code in asm info parsing
Now that all 3 supported filetypes have length information included,
we don't need to check for N50 or total length to be null (the
only assembly info field we need to check for its null status is
the assembly G/C content field).

I plan to remove some more unnecessary code from xdot2cy.js shortly.
This is a small start to that.