Update demo (fix #121) & add upload functionality
Pretty happy with the way the demo works now. Might want to reduce
horizontal space taken up by a bit (right now the controls bar just
barely fits on my screen, so on smaller screens it's going to
break lines), but aside from that this seems to be functioning fine.