Fix unserializable data issue: support JSON #12
Turns out the bug was on my end: collections of nodes are evidently
unserializable in JSON, which prevented the call to
JSON.stringify(cy.json()) from working. This data is now stored in
scratch, and I'm looking into using less scratch/data stuff in
general (a lot of this we can deduce from the elements themselves,
without extra space requirements of storing lots of stuff).