Remove outdated tool_design/ directory
Most of the stuff in it was pretty out-of-date by this point.
Furthermore, the code is mostly well-commented enough to serve
as alright self-documentation, to a point.

When we work on releasing/publishing this, I can use the tool_design
directory (it's saved on my system) somewhat as a basis for writing
about the tool, although I've written other stuff since the
tool_design directory's contents that is more recent, more detailed,
and more informative.