Add selected cluster info (#130)
Still need to allow the user to export DNA of selected clusters.
See added code for comments re: implementing that; should be
doable, but might require a bit of refactoring to be efficient.

Also, there's a silly bug in the "selected elements" button which
I'll fix in a few minutes.