Fix collapsed cluster scaling re: new scaling
It's gotten sort of complicated due to the dichotomy of width/height
parameters -- I should really take the time to unify those into
one thing, since it's pretty clear by now that we're sticking with
identical widths/heights.

In any case, this works well and as intended and I added some more
documentation. We'll probably want to eventually precompute this
stuff in collate.py (no reason not to), but for now it works fine.