Add "bb_enforcing" nodes: iterative drawing (#66)
This small addition ensures that clusters, when "iteratively"
drawn before other elements, are drawn in their approximately
correct positions relative to the viewport.

There's a better explanation of this within the comments of #66.

One more slight issue to tidy up with iterative drawing and then
this issue should be done.