Reconcile demo and local viewer interfaces
This will make it a lot easier to update the demo in the future.
The only downside is that the local version has the "Demo .db"
dialog enabled, which is kind of a problem (that just causes errors
since we don't store .dbs locally in the viewer/ folder)
but not a huge issue by any means (and one we can easily rectify
when actually releasing the tool).