Add dynamic+collapsible selected info: close #144
This looks nice, and it works well with the process of exploring the
graph (seeing the numbers indicating different types of selected
elements change is somewhat helpful).

One thing here: I changed things so that contigs (nodes in LastGraph
and GFA assembly graphs) have their length indicated in nt, while
scaffolds (nodes in BAMBUS 3 GML assembly graphs) have their
length indicated in bp. I'm fairly sure this is correct, but I'm
leaving #107 open for the time being so I can 1) add this to
the README and 2) double-check that this is, in fact, correct.

There's a lot of issues we could focus on next. #145 and #142, #143
(and once #142 and #143 are done, #68) are all fairly simple and
shouldn't take too long (knock on wood); #71, #80, and #90 are all
super high priority and should be done as soon as possible,
although they'll take a fairly substantial amount of work.