Fix conn. cmp. rank validity checking
Now, only valid integers are accepted. I also tried to make the
decrCompRank() and incrCompRank() code a bit more efficient, and
make its behavior as "natural" when using as possible -- hopefully
that comes across.

(I might've made the code a bit ugly, but it should be somewhat
more efficient than what we had before. Not sure if that's the
greatest tradeoff.)