Use Btstrp modal(asm info): close #143, close #68
As of this commit, we have now officially switched from jQuery UI
to Bootstrap. There should be a bit less overhead on the application
now, and it looks a lot nicer than it did before (although a lot of
the renovations are independent of jQuery UI/Bootstrap).

After this I'd like to switch focus to some of the more concrete
aspects of the application (i.e. ...not display stuff). This is a
good step toward getting the application polished for release,
though.