Use pygraphviz, rm dir "requirement": close #109
pygraphviz is now used to lay out the graph and get its .xdot
layout, which means that "intermediate" .gv/.xdot files don't have
to be created. (The -pg and -px options are still available, of
course.) Per #28, we could definitely use pygraphviz to parse .xdot
layout, and I plan to work on that right after I finish typing this
commit message (...and after I get dinner).

One noteworthy change -- now, if no -d argument is passed, then
files are just generated in the current working directory. I actually
really like this behavior, and I think it results in a much more
natural user interface than forcing the user to create a directory
every time they run the script.

I also updated the README with more specific and up-to-date
information on the behavior of collate.py. To eliminate redundancies
and potential inconsistencies, I just moved all documentation from
the header of collate.py to the README.