Use pygraphviz to record xdot info (close #28)
That was certainly an odyssey. See the issue thread for a summary
of the steps I walked through on the way (and a glimpse into my
sanity while working on the problem :-).

To summarize things: basically, the only concession I had to make
here was an inability to record the bounding box of each component.
That actually isn't really a major problem for the viewer
application, which surprised me: the reason for this is, regardless
of the bounding box, each graph element within a component will have
an identical offset from its "correct" position, and since we fit the
component elements to the screen this doesn't make a difference.

However, having an accurate bounding box is still useful for
iterative graph drawing (esp. when there aren't a lot of node
groups). So I'll see if I can get an approximation of the bounding
box included in the code soon.