Add save_aux_file for .gv/.xdot files (#102)
If we run into .gv / .xdot files that cannot be saved for one
reason or another (they're directories, say, or they already exist
and -w isn't set), then it doesn't make sense to just halt the
program with an error. What save_aux_file does is just print a
warning to the user that the aux file couldn't be saved, and continue
on. (Of course with .db files it's much more serious that they
be saved, but with .xdot/.gv files it isn't necessary thanks to
pygraphviz.)

I should probably put this in the README soon also.