Add cluster layout + width/height detection: #80
This necessitated a few changes in the surrounding code. Also the
width/height obtained here aren't really used anywhere yet (but
will be when we lay out the entire component graph using node
groups as just rectangle nodes).

I'd say we're about halfway done with Part 1 as detailed in the
"pseudocode" in #80. Hopefully I can make some faster progress
on this tomorrow.