Implement "atomic" node group layout (close #80)
This results in much more "linear"-looking graphs, at least with the
test assembly graphs I've tried. I'm pretty happy with how this
turned out [although implementing it was a bit involved :-)].

Edges to/from nodes that were within node groups look a bit off, so
I'm going to look into adding extra control points based on the
distance between the head-/tail-port of a rectangular node and the
actual head-/tail-port of the source/target node(s) in question.
I guess that gets filed under #56 for now (might create a separate
issue later).

We should probably also look into determining the effects (if any)
this has on layout times for large graphs (e.g. component 1 of the
Shakya assembly graph). If it causes a significant increase in layout
time, we should consider making this configurable for the user.

One last thing: I noticed on a few small tests that occasionally
we'll get a warning from graphviz that the graph size is too small
for a label. This is due to us using fixedsize=true for nodes, and
these warnings are entirely fine -- however, it might be worth
suppressing these warnings if possible so as to not make it seem like
things are going wrong when they aren't (although the warnings aren't
really horrible or anything, so that's not important).