Scale progbar update freq to cmp size: fix #151
The scaling applies to whatever's being drawn, so it adjusts as
different components/etc. are drawn. When we implement the ability
to draw multiple components at once (see #93), we can adjust this
to scale accordingly to the increased size.

This results in a pretty huge speedup, which is great! I think the
slightly choppy progress bar animations are more than an alright
cost, and we can probably look into making that a bit smoother
in the future (but that's super low priority relative to tons of
other stuff).