Use os.open for aux files; fix output msgs a bit
It looks like I accidentally reverted from using os.open() (with the
properly set flags, based on whether or not -w is set) when I
switched to using PyGraphviz instead of just manually calling
GraphViz a while ago (in commit 1b68572, to be exact). We now use
os.open() when saving auxiliary files (which is the one time we
actually directly write to a file from our side of things), so as
to ensure that no data is lost/overwritten even in the case of a
race condition. (See the docstring of check_file_existence() for
further details.)

I also added in some simple logic to save_aux_file() to return
a boolean depending on when or not an error message was printed when
trying to save an auxiliary file -- we use this to make the
formatting of error messages from the command line a bit nicer, with
no unnecessary line breaks in between error messages (in case, say,
both -pg and -px are set and a given component encounters errors when
trying to draw both its .gv and .xdot files).

I think this should resolve all the stuff having to deal with #26.
Most of what this addresses are very marginal, corner-case-ish errors
but I figure it's good to take care of that stuff in advance when
possible.