Add Cytoscape.js layout testing (close #100)
It's through an actual UI interface, so it's fairly simple to play
around with. I also added in support for the cose-bilkent layout
plugin, which works nicely with compound graphs but seems to be
decently slower than the normal cose builtin layout.

I don't know if we're going to keep this feature around for the
release of the tool, but it's good for testing in the meantime.