Only display non-compound node labels
A lot of compound nodes can have large labels (since they can
contain a lot of nodes), which can be expensive to render and mess
up fitting the graph/etc.

Also, a node group should probably be characterized mostly by its
1) structure and 2) child nodes, so having a node group ID displayed
isn't very helpful.

I guess these IDs are still displayed in the "Selected Cluster Info"
table, which could be kind of long... but I'm not sure of a better
way to handle that at present. Maybe display the "source" and
"sink" node IDs of the node group in question?
-> Although that wouldn't work for frayed ropes and might be tricky
for linear cycles, though.