Add SPQR bubble detection (#71)
This isn't quite done yet -- there seems to be a bug somewhere that's
caused really large bubbles to be generated in certain scenarios.
Once that's handled, I think this should be taken care of?
Committing this anyway because this is a lot of code, and I'd like
to have it stored on GitHub even if it does have that bug (the
previous bubble detection code is commented out, so it should be
easily revertible if we decide to do that instead).

I'm not sure if we'd want to actually draw the entire SPQR tree
or just highlight bubbles, like we do here. If we want to draw the
entire tree then I might need to write some more code to interface
with OGDF to do that?

To the spirit of compressing the graph, we could later also collapse
all identified node groups by default in the viewer (using the
"backfilled" node group sizes to have correct edges).