Display Bambus edge props. in viewer: close #153
Also, the edge header table structure now changes based on the
type of assembly graph file you load. We're almost finished with #135
as well -- just need to implement that for nodes, then that issue
should be done as well.

This was a pretty large commit, so it contained a few other changes.
Multiplicity in .dbs now contains bundle size for GML files and
multiplicity for LastGraph files -- this means we don't have a
redundant column (since the two columns were mutually exclusive).
We format the edge header in the viewer for multiplicity as
"B. size" for GML files, so the user can tell what's going on.

A few code tweaks here also, to make things easier to understand/to
modify in the future. I removed the .viewable class from the viewer
CSS, since that was kind of useless (everything contained in it was
a default CSS property). I also made changing up the statements to
add things to the .db file in python easier in the future, by
declaring variables for each statement (and then reusing those
statements).

I think that's mostly it. Sorry if this commit message is hard to
read; I'm a bit tired.