Don't display cluster IDs in viewer application
For really large node groups (some SPQR-detected bubbles, some large
chains, etc) the IDs can get large enough that they don't really
yield any important information to the user and cause the info table
to force horizontal scrolling in the controls div. So I think this
is an alright approach.

I was thinking of just displaying the source/sink IDs of a node
group in the info table, but that doesn't really hold for
some SPQR-detected bubbles, for cyclic chains, or for frayed ropes.