Adjust python code to parse Bambus' GFA output
adjustments for extracting node IDs from node names (specific
to Bambus' node names; will have to adjust to be more general in
future) and for extracting length info from sequence lines without
DNA (using the LN:i: property).

Note that no DNA means no G/C content stats available, which means
that in the viewer right now these files are displayed as having
a G/C content of 0%. As a TODO, adjust viewer to handle this
(displaying "N/A" should be okay for the time being -- since that
peculiar lack of data is not specific to the filetype).