Add test "single graph" functionality+misc changes
See #10 for the full story here. Basically, single graphs might
not be that useful for the sort of analysis we want to do using this
tool -- in Bambus 3's output, contigs already have an orientation
while in LastGraph/GFA output that isn't a guarantee.

Our current behavior is to auto-draw double graphs for LastGraph and
GFA files, and auto-draw just the graph structure (I don't know if
you'd call it a "single graph") for Bambus 3 GML files.

Using explicit "double graphs" for GML files just results in 2x the
connected components with no real added information (since nodes
already have an orientation -- all nodes are "positive", so reverse
complementing all nodes in the graph just creates new connections
between the new negative nodes).

Using "single graphs" for
LastGraph/GFA files in general, though, does not keep the same level
of information -- there exists the possibility for a given node and
its RC to be in the same connected component, and nodes + edges
must be treated as directionless. The effect of this is that
analyzing the structural properties of the graph becomes a decent
amount harder (not to mention that just laying out the graph
hierarchically depends on the graph being a digraph). Essentially,
we could include this functionality but it wouldn't really tie in
with any of the Bambus 3 stuff or with most of the other features
AsmViz has right now.

As long as we're explicit in the README about what sorts of drawings
are produced, I think keeping the current policy (default structure
for Bambus 3 GML files, double graphs for other filetypes) should be
alright. The tool kind of focuses on scaffold graphs, anyway.