Make node polygon rotation more efficient (#115)
Instead of each node having a data(polypts) property that is rotated
in rotateNode(), all possible node polygon rotations are defined in
advance and are adjusted via just class changes.

Now that rotation is basically fixed (it's at least a lot more
efficient than it used to be a few hours ago), I can get to work on
implementing #136.