Give B/R/Y node groups collapsed shapes: fix #136
(That is, Bubbles/Frayed Ropes/Cyclic chains.)

These shapes make looking through a collapsed graph a bit easier,
and provide a sort of visual cue as to the contents of these node
groups. Also, they look fancy.

Worth noting that these special collapsed shapes do apply while the
graph is being iteratively drawn (so the user will see hexagonal
bubbles, etc. stretched out to their uncollapsed size) -- I think
this is ok for now, and actually helps give the user some information
about the overall graph structure in advance, although we may want
to consider changing that in the future to the previous thing (so
that iteratively drawing only draws rectangular node groups).

This commit was the result of overhauling the rotation system today
to make it a lot more sane. Adding similar changes like this in the
future should be a lot simpler now.