Save scaff. info with line-by-line approach #168
Also figured out another thing that was causing the Judson
stuff to fail (a small issue -- just an extra character in the file
that was throwing my code off).

I don't consider this "done" yet. The process of checking if contigs
are in the current component is super computationally expensive for
large graphs and/or large scaffold files -- to the point where it's
a lot faster to just not filter scaffolds and just display all of
them in the viewer UI than only to display filtered scaffolds.

I've thought of a few ideas re: potentially speeding up filtering.
One approach: save invalid scaffold IDs so the filtering process
involves O(|scaffolds|) filtering operations. This would make it
about 3x faster (assuming average of 3 contigs per scaffold),
which is decent but not what I'd consider good. I'll think of
some things.