Remove scaling test stuff from repo
Since #175 is done.

I still have this saved on my machine, and it's synced via Dropbox,
so this should still be ready if we want to try it out again.

(Also I sent these files as an email attachment at one point so
there's going to be at least two backups of this stuff, so I'm not
too worried about losing it)