Make AGP files consistent across cmps (fix #174)
The solution to this required a bit of code restructuring. It takes
advantage of the fact that parseDBcomponents() must be called every
time a new assembly is loaded -- this allows us to clear loaded
scaffold info in that function, but not when drawing a new component.

Also, turns out that {} === {} returns false in JavaScript. huh.
(spent a few minutes trying to figure out what was going in my code
-- hence the creation of isObjectNonempty().)