Remove graph rotation from viewer interface
Based on feedback, this isn't a very useful feature at present.
(Also, it messes up with the new node width/height scaling stuff.)

I kept all of the code for this around in the files, so if I need
to add it back in the future that shouldn't be too difficult.