Implement sane zooming limits: resolve #166
Minimum zoom is based on the zoom level adopted by Cytoscape.js when
the entire graph has been fit to the screen. Maximum zoom is set
automatically -- I looked at a bunch of graphs, large and small,
and found that "9" was always an upper bound on the zoom level of
fitting the screen to a single node. So we just use that.

Ideally we'd set max zoom more precisely, but due to differing node
sizes I'm not sure how we'd do that. For now this should be okay --
it's a lot harder for users to get lost via zooming now, at least.