Fix "bounding box enforcing" nodes (#148)
Turns out that a node height or width of zero causes Cytoscape.js
to just not draw these nodes -- which meant that they were not used
in fitting the graph, which in turn meant that these nodes' purpose
of impacting the fitting of the graph was not being fulfilled.

Making these nodes invisible and with a width/height of 1 seems to
work.

There's still some accuracy problems in cluster drawings (#148), but
after this update those problems are so minor that fixing them is a
very low priority issue.