Draw edges to/from interior nodes: close #150
This isn't a panacea for complex edges being labelled as invalid
bezier curves by Cytoscape.js. Interestingly enough, in the first
component of the E. coli assembly graph a few edges seemed to
actually have become "badbezier" edges due to this change.

However, I think those issues are things that will have to be
addressed in #56.